refactor(solve): prove compact GPU initialization shape
