+++ crates/rumoca-phase-solve/src/lib.rs
@@ -8,9 +8,8 @@
 //! The DAE tree-walk interpreter (`eval`, `dual`, `sim_float`, `statement`) lives
 //! in `rumoca-eval-dae`.
 //!
-//! SPEC_0021 file-size exception: phase-solve facade still owns exports,
-//! compatibility tests, and lowering integration. split plan: keep moving
-//! tests and integration-only helpers into focused modules.
+//! The facade owns exports and lowering integration; GPU initialization and
+//! projection planning live in focused modules to keep phase boundaries legible.
 
 use std::collections::{BTreeMap, BTreeSet, HashSet};
 
@@ -33,12 +32,14 @@ mod discrete_pre_modes;
 mod dummy_derivative;
 mod dynamic_events;
 mod event_actions;
+mod gpu_initialization;
 mod implicit_rhs;
 mod initial_values;
 pub mod layout;
 pub mod lower;
 mod observation_refresh;
 mod path_utils;
+mod projection_plan;
 mod projection_suffix;
 mod residual_compute_block;
 mod runtime_assignments;
@@ -70,6 +71,9 @@ use continuous_row_targets::{
 use discrete_pre_modes::discrete_pre_mode_for_equation;
 #[cfg(test)]
 pub(crate) use discrete_pre_modes::expression_contains_event_entry_pre_operator;
+use gpu_initialization::lower_gpu_initialization_system;
+#[cfg(test)]
+use gpu_initialization::{append_gpu_corner_strides, gpu_corner_cell_index};
 #[cfg(test)]
 use implicit_rhs::zero_rhs_row;
 use implicit_rhs::{
@@ -88,6 +92,7 @@ use lower::{
     normalized_discrete_update_equations,
 };
 use observation_refresh::lower_discrete_observation_refresh;
+use projection_plan::*;
 use runtime_assignments::{
     lower_runtime_assignment_targets, runtime_assignment_equation, runtime_assignment_equations,
     runtime_tail_update_names, static_runtime_tail_equation,
@@ -770,539 +775,6 @@ fn lower_initialization_system(
     })
 }
 
-/// GPU preparation deliberately accepts only direct, regular initial families.
-/// It builds one base row plus one corner per binder, never a vector of scalar
-/// rows. Runtime initialization keeps its complete scalar/general path above.
-fn lower_gpu_initialization_system(
-    dae_model: &dae::Dae,
-    layout: &solve::VarLayout,
-) -> Result<solve::InitializationSolveSystem, LowerError> {
-    if dae_model.initialization.equations.is_empty() {
-        return Ok(solve::InitializationSolveSystem::default());
-    }
-    let mut expected = 0usize;
-    let mut nodes = Vec::new();
-    let mut families = Vec::new();
-    let mut residual_start = 0usize;
-    for family in &dae_model.initialization.structured_equations {
-        let Some(_regular) = family.regular.as_ref() else {
-            return Err(LowerError::Unsupported {
-                reason: "GPU initial projection requires a regular structured initial family"
-                    .to_string(),
-            });
-        };
-        let Some(template) = family.template.as_ref() else {
-            return Err(LowerError::Unsupported {
-                reason: "GPU initial projection requires a structured initial template".to_string(),
-            });
-        };
-        let Some(body_count) = family.common_iteration_equation_count() else {
-            return Err(LowerError::Unsupported {
-                reason:
-                    "GPU initial projection requires a nonempty uniform structured initial family"
-                        .to_string(),
-            });
-        };
-        if body_count == 0 || template.body.len() != body_count {
-            return Err(LowerError::Unsupported {
-                reason: "GPU initial projection requires one uniform template body per family cell"
-                    .to_string(),
-            });
-        }
-        let cells = family
-            .domain
-            .scalar_count()
-            .map_err(|error| LowerError::contract_violation(error.to_string(), family.span))?;
-        expected = expected
-            .checked_add(cells.checked_mul(body_count).ok_or_else(|| {
-                LowerError::contract_violation("GPU initial family size overflow", family.span)
-            })?)
-            .ok_or_else(|| {
-                LowerError::contract_violation("GPU initial residual size overflow", family.span)
-            })?;
-        for position in 0..body_count {
-            let direct = lower_gpu_direct_family(
-                dae_model,
-                layout,
-                family,
-                position,
-                body_count,
-                residual_start,
-            )?;
-            residual_start = residual_start.checked_add(cells).ok_or_else(|| {
-                LowerError::contract_violation("GPU initial residual range overflow", family.span)
-            })?;
-            nodes.push(direct.residual);
-            let node_index = nodes.len() - 1;
-            let direct = solve::InitializationDirectFamily {
-                node_index,
-                targets: direct.targets,
-                residual_sign: direct.residual_sign,
-                span: direct.span,
-            };
-            families.push(direct);
-        }
-    }
-    if dae_model.initialization.equation_provenance.len()
-        != dae_model.initialization.equations.len()
-    {
-        return Err(LowerError::Unsupported {
-            reason: "GPU initial projection requires typed provenance for every initial equation"
-                .to_string(),
-        });
-    }
-    let required_user_initial_rows = dae_model
-        .initialization
-        .equations
-        .iter()
-        .zip(&dae_model.initialization.equation_provenance)
-        .filter(|(_, provenance)| **provenance != dae::InitializationEquationProvenance::FixedStart)
-        .map(|(equation, _)| equation)
-        .try_fold(0usize, |total, equation| {
-            total
-                .checked_add(equation.scalar_count.max(1))
-                .ok_or_else(|| {
-                    LowerError::contract_violation(
-                        "GPU initial user-row count overflow",
-                        equation.span,
-                    )
-                })
-        })?;
-    if expected != required_user_initial_rows {
-        return Err(LowerError::Unsupported { reason: "GPU initial projection requires complete structured coverage; mixed or nonstructured initial rows are unsupported".to_string() });
-    }
-    require_complete_gpu_initial_target_coverage(dae_model, layout, &families)?;
-    Ok(solve::InitializationSolveSystem {
-        residual: solve::ComputeBlock { nodes },
-        direct_families: families,
-        ..Default::default()
-    })
-}
-
-fn require_complete_gpu_initial_target_coverage(
-    dae_model: &dae::Dae,
-    layout: &solve::VarLayout,
-    families: &[solve::InitializationDirectFamily],
-) -> Result<(), LowerError> {
-    let mut covered = vec![false; layout.y_scalars()];
-    for (structured, direct) in dae_model
-        .initialization
-        .structured_equations
-        .iter()
-        .flat_map(|structured| {
-            (0..structured.common_iteration_equation_count().unwrap_or(0)).map(move |_| structured)
-        })
-        .zip(families)
-    {
-        for index in direct
-            .targets
-            .output_indices(&structured.domain)
-            .map_err(|error| LowerError::contract_violation(format!("{error:?}"), direct.span))?
-        {
-            let Some(slot) = covered.get_mut(index) else {
-                return Err(LowerError::contract_violation(
-                    "GPU initial target is outside the solver Y vector",
-                    direct.span,
-                ));
-            };
-            *slot = true;
-        }
-    }
-    for (equation, provenance) in dae_model
-        .initialization
-        .equations
-        .iter()
-        .zip(&dae_model.initialization.equation_provenance)
-    {
-        if *provenance != dae::InitializationEquationProvenance::FixedStart {
-            continue;
-        }
-        let targets = lower_continuous_row_targets_for_equation(
-            dae_model,
-            equation,
-            layout,
-            equation.scalar_count.max(1),
-        )?;
-        for target in targets {
-            let Some(solve::ScalarSlot::Y { index, .. }) = target else {
-                return Err(LowerError::contract_violation(
-                    "GPU fixed-start initialization requires a Y target",
-                    equation.span,
-                ));
-            };
-            let Some(slot) = covered.get_mut(index) else {
-                return Err(LowerError::contract_violation(
-                    "GPU fixed-start target is outside the solver Y vector",
-                    equation.span,
-                ));
-            };
-            *slot = true;
-        }
-    }
-    if covered.iter().any(|covered| !covered) {
-        return Err(LowerError::Unsupported {
-            reason: "GPU initial projection requires the union of user equations and fixed starts to cover every solver Y slot".to_string(),
-        });
-    }
-    Ok(())
-}
-
-fn lower_gpu_direct_family(
-    dae_model: &dae::Dae,
-    layout: &solve::VarLayout,
-    family: &dae::StructuredEquationFamily,
-    position: usize,
-    body_count: usize,
-    residual_start: usize,
-) -> Result<GpuLoweredDirectFamily, LowerError> {
-    let base_index = family
-        .first_equation_index
-        .checked_add(position)
-        .ok_or_else(|| {
-            LowerError::contract_violation("GPU initial base equation index overflow", family.span)
-        })?;
-    let base_equation = dae_model
-        .initialization
-        .equations
-        .get(base_index)
-        .ok_or_else(|| {
-            LowerError::contract_violation("GPU initial base equation is missing", family.span)
-        })?;
-    let base_ops = lower_initial_residual_cell(
-        dae_model,
-        layout,
-        dae_model.continuous.equations.len() + base_index,
-        base_equation,
-    )?;
-    let base_target = direct_initial_target(dae_model, layout, base_equation, family.span)?;
-    let sign = direct_initial_assignment_sign(&base_ops, base_target).ok_or_else(|| {
-        LowerError::Unsupported {
-            reason: "GPU initial projection requires a direct target-minus-rhs structured row"
-                .to_string(),
-        }
-    })?;
-    let strides = lower_gpu_direct_family_strides(
-        dae_model,
-        layout,
-        family,
-        position,
-        body_count,
-        &base_ops,
-        base_target,
-    )?;
-    Ok(GpuLoweredDirectFamily {
-        residual: solve::ComputeNode::Map {
-            domain: family.domain.clone(),
-            output_map: solve::TensorOutputMap::dense_contiguous(residual_start, &family.domain)
-                .map_err(|error| {
-                    LowerError::contract_violation(format!("{error:?}"), family.span)
-                })?,
-            base_ops,
-            load_strides: strides.loads,
-            const_strides: strides.constants,
-            metadata: solve::TensorNodeMetadata::default(),
-            span: family.span,
-        },
-        targets: solve::TensorOutputMap {
-            start: base_target,
-            strides: strides.targets,
-        },
-        residual_sign: sign,
-        span: family.span,
-    })
-}
-
-struct GpuLoweredDirectFamily {
-    residual: solve::ComputeNode,
-    targets: solve::TensorOutputMap,
-    residual_sign: i8,
-    span: rumoca_core::Span,
-}
-
-struct GpuDirectFamilyStrides {
-    loads: Vec<solve::AffineStencilLoadStride>,
-    constants: Vec<solve::AffineStencilConstStride>,
-    targets: Vec<solve::AffineStencilIndexStrideTerm>,
-}
-
-fn lower_gpu_direct_family_strides(
-    dae_model: &dae::Dae,
-    layout: &solve::VarLayout,
-    family: &dae::StructuredEquationFamily,
-    position: usize,
-    body_count: usize,
-    base_ops: &[solve::LinearOp],
-    base_target: usize,
-) -> Result<GpuDirectFamilyStrides, LowerError> {
-    let mut load_strides = Vec::new();
-    let mut const_strides = Vec::new();
-    let mut target_strides = Vec::new();
-    for dimension in 0..family.domain.binders.len() {
-        let corner_index = gpu_direct_family_corner_index(family, position, body_count, dimension)?;
-        let corner_equation = dae_model
-            .initialization
-            .equations
-            .get(corner_index)
-            .ok_or_else(|| {
-                LowerError::contract_violation(
-                    "GPU initial corner equation is missing",
-                    family.span,
-                )
-            })?;
-        let corner_ops = lower_initial_residual_cell(
-            dae_model,
-            layout,
-            dae_model.continuous.equations.len() + corner_index,
-            corner_equation,
-        )?;
-        let corner_target = direct_initial_target(dae_model, layout, corner_equation, family.span)?;
-        target_strides.push(solve::AffineStencilIndexStrideTerm {
-            dimension,
-            stride: gpu_initial_stride(corner_target, base_target, family.span, "target")?,
-        });
-        append_gpu_corner_strides(
-            base_ops,
-            &corner_ops,
-            dimension,
-            &mut load_strides,
-            &mut const_strides,
-            family.span,
-        )?;
-    }
-    Ok(GpuDirectFamilyStrides {
-        loads: load_strides,
-        constants: const_strides,
-        targets: target_strides,
-    })
-}
-
-fn gpu_direct_family_corner_index(
-    family: &dae::StructuredEquationFamily,
-    position: usize,
-    body_count: usize,
-    dimension: usize,
-) -> Result<usize, LowerError> {
-    let corner_cell = gpu_corner_cell_index(&family.domain, dimension, family.span)?;
-    family
-        .first_equation_index
-        .checked_add(corner_cell.checked_mul(body_count).ok_or_else(|| {
-            LowerError::contract_violation(
-                "GPU initial corner equation index overflow",
-                family.span,
-            )
-        })?)
-        .and_then(|value| value.checked_add(position))
-        .ok_or_else(|| {
-            LowerError::contract_violation(
-                "GPU initial corner equation index overflow",
-                family.span,
-            )
-        })
-}
-
-fn gpu_initial_stride(
-    corner: usize,
-    base: usize,
-    span: rumoca_core::Span,
-    kind: &'static str,
-) -> Result<isize, LowerError> {
-    isize::try_from(corner)
-        .ok()
-        .and_then(|value| value.checked_sub(isize::try_from(base).ok()?))
-        .ok_or_else(|| {
-            LowerError::contract_violation(format!("GPU initial {kind} stride overflows"), span)
-        })
-}
-
-fn direct_initial_target(
-    dae_model: &dae::Dae,
-    layout: &solve::VarLayout,
-    equation: &dae::Equation,
-    span: rumoca_core::Span,
-) -> Result<usize, LowerError> {
-    let targets = lower_continuous_row_targets_for_equation(dae_model, equation, layout, 1)?;
-    match targets.as_slice() {
-        [Some(solve::ScalarSlot::Y { index, .. })] => Ok(*index),
-        _ => Err(LowerError::contract_violation(
-            "GPU initial projection requires one Y target per direct family row",
-            span,
-        )),
-    }
-}
-
-fn gpu_corner_cell_index(
-    domain: &rumoca_core::StructuredIndexDomain,
-    dimension: usize,
-    span: rumoca_core::Span,
-) -> Result<usize, LowerError> {
-    let binder = domain.binders.get(dimension).ok_or_else(|| {
-        LowerError::contract_violation("GPU initial corner dimension is missing", span)
-    })?;
-    if gpu_binder_value_count(binder, span)? < 2 {
-        return Err(LowerError::Unsupported {
-            reason: "GPU initial projection requires a non-degenerate structured binder"
-                .to_string(),
-        });
-    }
-    domain.binders[dimension + 1..]
-        .iter()
-        .try_fold(1usize, |stride, later| {
-            let count = gpu_binder_value_count(later, span)?;
-            stride.checked_mul(count).ok_or_else(|| {
-                LowerError::contract_violation("GPU initial corner stride overflow", span)
-            })
-        })
-}
-
-fn gpu_binder_value_count(
-    binder: &rumoca_core::StructuredIndexBinder,
-    span: rumoca_core::Span,
-) -> Result<usize, LowerError> {
-    if binder.step == 0 {
-        return Err(LowerError::contract_violation(
-            "GPU initial binder step must be nonzero",
-            span,
-        ));
-    }
-    let distance = if binder.step > 0 {
-        binder.upper.checked_sub(binder.lower)
-    } else {
-        binder.lower.checked_sub(binder.upper)
-    }
-    .ok_or_else(|| LowerError::contract_violation("GPU initial binder bounds are invalid", span))?;
-    let step = binder.step.unsigned_abs();
-    let count = distance
-        .checked_div(i64::try_from(step).map_err(|_| {
-            LowerError::contract_violation("GPU initial binder step overflow", span)
-        })?)
-        .and_then(|value| value.checked_add(1))
-        .ok_or_else(|| LowerError::contract_violation("GPU initial binder count overflow", span))?;
-    usize::try_from(count).map_err(|_| {
-        LowerError::contract_violation("GPU initial binder count exceeds host range", span)
-    })
-}
-
-fn append_gpu_corner_strides(
-    base: &[solve::LinearOp],
-    corner: &[solve::LinearOp],
-    dimension: usize,
-    load_strides: &mut Vec<solve::AffineStencilLoadStride>,
-    const_strides: &mut Vec<solve::AffineStencilConstStride>,
-    span: rumoca_core::Span,
-) -> Result<(), LowerError> {
-    if base.len() != corner.len() {
-        return Err(LowerError::Unsupported {
-            reason: "GPU initial projection requires identical direct-family operation shapes"
-                .to_string(),
-        });
-    }
-    for (op_position, (base_op, corner_op)) in base.iter().zip(corner).enumerate() {
-        match (base_op, corner_op) {
-            (
-                solve::LinearOp::LoadY { index: base, .. },
-                solve::LinearOp::LoadY { index: corner, .. },
-            ) => {
-                let stride = isize::try_from(*corner)
-                    .ok()
-                    .and_then(|value| value.checked_sub(isize::try_from(*base).ok()?))
-                    .ok_or_else(|| {
-                        LowerError::contract_violation("GPU initial Y stride overflows", span)
-                    })?;
-                if stride != 0 {
-                    load_strides.push(solve::AffineStencilLoadStride {
-                        op_position,
-                        terms: vec![solve::AffineStencilIndexStrideTerm { dimension, stride }],
-                    });
-                }
-            }
-            (
-                solve::LinearOp::LoadP { index: base, .. },
-                solve::LinearOp::LoadP { index: corner, .. },
-            ) => {
-                let stride = isize::try_from(*corner)
-                    .ok()
-                    .and_then(|value| value.checked_sub(isize::try_from(*base).ok()?))
-                    .ok_or_else(|| {
-                        LowerError::contract_violation("GPU initial P stride overflows", span)
-                    })?;
-                if stride != 0 {
-                    load_strides.push(solve::AffineStencilLoadStride {
-                        op_position,
-                        terms: vec![solve::AffineStencilIndexStrideTerm { dimension, stride }],
-                    });
-                }
-            }
-            (
-                solve::LinearOp::Const { value: base, .. },
-                solve::LinearOp::Const { value: corner, .. },
-            ) => {
-                let stride = corner - base;
-                if !stride.is_finite() {
-                    return Err(LowerError::contract_violation(
-                        "GPU initial constant stride is not finite",
-                        span,
-                    ));
-                }
-                if stride != 0.0 {
-                    const_strides.push(solve::AffineStencilConstStride {
-                        op_position,
-                        terms: vec![solve::AffineStencilConstStrideTerm { dimension, stride }],
-                    });
-                }
-            }
-            (
-                solve::LinearOp::LoadY { .. }
-                | solve::LinearOp::LoadP { .. }
-                | solve::LinearOp::Const { .. },
-                _,
-            ) => {
-                return Err(LowerError::Unsupported {
-                    reason: "GPU initial projection requires uniform direct-family access kinds"
-                        .to_string(),
-                });
-            }
-            _ => {}
-        }
-    }
-    Ok(())
-}
-
-fn direct_initial_assignment_sign(ops: &[solve::LinearOp], target_index: usize) -> Option<i8> {
-    let solve::LinearOp::StoreOutput { src } = ops.last()? else {
-        return None;
-    };
-    let solve::LinearOp::Binary {
-        op: solve::BinaryOp::Sub,
-        lhs,
-        rhs,
-        dst,
-    } = ops
-        .iter()
-        .find(|op| matches!(op, solve::LinearOp::Binary { dst, .. } if dst == src))?
-    else {
-        return None;
-    };
-    let target_loads = ops
-        .iter()
-        .filter_map(|op| match op {
-            solve::LinearOp::LoadY { dst, index } if *index == target_index => Some(*dst),
-            solve::LinearOp::LoadY { .. } => Some(u32::MAX),
-            _ => None,
-        })
-        .collect::<Vec<_>>();
-    if target_loads.len() != 1 || target_loads[0] == u32::MAX || *dst != *src {
-        return None;
-    }
-    let residual_sign = if target_loads[0] == *lhs {
-        1
-    } else if target_loads[0] == *rhs {
-        -1
-    } else {
-        return None;
-    };
-    Some(residual_sign)
-}
-
 fn lower_initialization_updates_only(
     dae_model: &dae::Dae,
     layout: &solve::VarLayout,
@@ -1533,812 +1005,6 @@ fn solver_residual_equation(
     Ok(!is_state_derivative_row
         && !static_runtime_tail_equation(dae_model, runtime_tail_updates, eq)?
         && runtime_assignment_equation(dae_model, runtime_tail_updates, eq)?.is_none())
-}
-
-fn lower_algebraic_projection_plan(
-    rows: &[Vec<solve::LinearOp>],
-    row_targets: &[Option<solve::ScalarSlot>],
-    state_scalar_count: usize,
-    solver_scalar_count: usize,
-    context_span: rumoca_core::Span,
-) -> Result<solve::AlgebraicProjectionPlan, LowerError> {
-    let projection_count = solver_scalar_count
-        .checked_sub(state_scalar_count)
-        .ok_or_else(|| {
-            lower_contract_violation(
-                "algebraic projection range starts after solver scalar count".to_string(),
-                context_span,
-            )
-        })?;
-    let mut projection_indices = lower_vec_with_capacity(
-        projection_count,
-        "algebraic projection index count",
-        context_span,
-    )?;
-    projection_indices.extend(state_scalar_count..solver_scalar_count);
-    lower_projection_plan(
-        rows,
-        row_targets,
-        &projection_indices,
-        state_scalar_count..solver_scalar_count,
-        ProjectionPlanPolicy {
-            include_explicit_row_targets: true,
-            require_complete_algebraic_coverage: true,
-        },
-        None,
-        context_span,
-    )
-}
-
-#[derive(Clone, Copy)]
-struct ProjectionPlanPolicy {
-    include_explicit_row_targets: bool,
-    require_complete_algebraic_coverage: bool,
-}
-
-// SPEC_0021: Exception - projection planning keeps explicit target,
-// initialization incidence, and identity-row ownership decisions together.
-#[allow(clippy::excessive_nesting)]
-fn lower_projection_plan(
-    rows: &[Vec<solve::LinearOp>],
-    row_targets: &[Option<solve::ScalarSlot>],
-    projection_indices: &[usize],
-    row_indices: std::ops::Range<usize>,
-    policy: ProjectionPlanPolicy,
-    implicit_incidence_rows: Option<&BTreeSet<usize>>,
-    context_span: rumoca_core::Span,
-) -> Result<solve::AlgebraicProjectionPlan, LowerError> {
-    let mut row_to_vars = BTreeMap::<usize, BTreeSet<usize>>::new();
-    let projection_set = projection_indices.iter().copied().collect::<BTreeSet<_>>();
-    let row_indices = row_indices.collect::<Vec<_>>();
-    let identity_projection_rows = row_indices
-        .iter()
-        .filter_map(|row_idx| {
-            identity_projection_y_index(rows.get(*row_idx)?.as_slice(), &projection_set)
-                .map(|y_idx| (y_idx, *row_idx))
-        })
-        .collect::<BTreeMap<_, _>>();
-
-    for &row_idx in &row_indices {
-        let mut y_indices = if implicit_incidence_rows.is_none_or(|rows| rows.contains(&row_idx)) {
-            collect_algebraic_y_indices_for_row(rows[row_idx].as_slice(), &projection_set)
-        } else {
-            BTreeSet::new()
-        };
-        let explicit_target = if let Some(solve::ScalarSlot::Y { index, .. }) =
-            row_targets.get(row_idx).copied().flatten()
-            && projection_set.contains(&index)
-        {
-            if policy.include_explicit_row_targets {
-                y_indices.insert(index);
-            } else {
-                y_indices.clear();
-                y_indices.insert(index);
-            }
-            true
-        } else {
-            false
-        };
-        if !explicit_target {
-            if let Some(index) =
-                identity_projection_y_index(rows[row_idx].as_slice(), &projection_set)
-            {
-                y_indices.clear();
-                y_indices.insert(index);
-            } else {
-                y_indices.retain(|index| {
-                    projection_index_not_claimed_by_identity(
-                        &identity_projection_rows,
-                        *index,
-                        row_idx,
-                    )
-                });
-            }
-        }
-        if y_indices.is_empty() {
-            continue;
-        }
-        row_to_vars.insert(row_idx, y_indices);
-    }
-
-    let projection_incidence = algebraic_projection_incidence(&row_to_vars, context_span)?;
-    let (blocks, dropped_equations) = projection_blt_blocks(&projection_incidence)?;
-    let mut blocks =
-        lower_blt_projection_blocks(&blocks, row_targets, &projection_incidence, context_span)?;
-    if policy.require_complete_algebraic_coverage {
-        blocks = retain_dropped_projection_rows(
-            blocks,
-            &dropped_equations,
-            &projection_incidence,
-            context_span,
-        )?;
-        validate_complete_algebraic_projection_plan(
-            &blocks,
-            rows,
-            &row_indices,
-            &projection_incidence,
-            context_span,
-        )?;
-    }
-    Ok(solve::AlgebraicProjectionPlan { blocks })
-}
-
-fn projection_index_not_claimed_by_identity(
-    identity_projection_rows: &BTreeMap<usize, usize>,
-    index: usize,
-    row_idx: usize,
-) -> bool {
-    identity_projection_rows
-        .get(&index)
-        .is_none_or(|identity_row| *identity_row == row_idx)
-}
-
-fn identity_projection_y_index(
-    row: &[solve::LinearOp],
-    projection_set: &BTreeSet<usize>,
-) -> Option<usize> {
-    let [
-        solve::LinearOp::LoadY {
-            dst: load_dst,
-            index,
-        },
-        solve::LinearOp::StoreOutput { src },
-    ] = row
-    else {
-        return None;
-    };
-    (*load_dst == *src && projection_set.contains(index)).then_some(*index)
-}
-
-fn projection_blt_blocks(
-    projection_incidence: &ProjectionIncidence,
-) -> Result<(Vec<BltBlock>, Vec<EquationRef>), LowerError> {
-    if projection_incidence.incidence.n_eq == 0 && projection_incidence.incidence.n_var == 0 {
-        return Ok((Vec::new(), Vec::new()));
-    }
-    let regular =
-        rumoca_phase_structural::maximum_regular_subsystem(&projection_incidence.incidence)
-            .map_err(|err| LowerError::Unsupported {
-                reason: format!("lower algebraic projection BLT: {err}"),
-            })?;
-    let blocks =
-        rumoca_phase_structural::build_blt_from_incidence(&regular.incidence).map_err(|err| {
-            LowerError::Unsupported {
-                reason: format!("lower algebraic projection BLT: {err}"),
-            }
-        })?;
-    Ok((blocks, regular.dropped_equations))
-}
-
-fn retain_dropped_projection_rows(
-    mut blocks: Vec<solve::AlgebraicProjectionBlock>,
-    dropped_equations: &[EquationRef],
-    projection_incidence: &ProjectionIncidence,
-    context_span: rumoca_core::Span,
-) -> Result<Vec<solve::AlgebraicProjectionBlock>, LowerError> {
-    for equation in dropped_equations {
-        let row_y_indices = projection_row_y_indices(equation.0, projection_incidence);
-        let mut retained = lower_vec_with_capacity(
-            blocks.len(),
-            "retained algebraic projection block count",
-            context_span,
-        )?;
-        let mut insertion_index = None;
-        let mut merged = None;
-        for block in blocks {
-            if block
-                .y_indices
-                .iter()
-                .any(|index| row_y_indices.contains(index))
-            {
-                insertion_index.get_or_insert(retained.len());
-                merged = Some(match merged {
-                    Some(previous) => combine_projection_blocks(previous, block, context_span)?,
-                    None => block,
-                });
-            } else {
-                retained.push(block);
-            }
-        }
-        if let (Some(mut block), Some(index)) = (merged, insertion_index) {
-            reserve_lower_capacity(
-                &mut block.rows,
-                1,
-                "retained algebraic projection row count",
-                context_span,
-            )?;
-            block.rows.push(equation.0);
-            block.rows.sort_unstable();
-            block.rows.dedup();
-            retained.insert(index, block);
-        }
-        blocks = retained;
-    }
-    Ok(blocks)
-}
-
-fn projection_row_y_indices(
-    row: usize,
-    projection_incidence: &ProjectionIncidence,
-) -> BTreeSet<usize> {
-    let Some(position) = projection_incidence
-        .incidence
-        .equation_refs
-        .iter()
-        .position(|equation| equation.0 == row)
-    else {
-        return BTreeSet::new();
-    };
-    projection_incidence.incidence.eq_unknowns[position]
-        .iter()
-        .filter_map(|unknown_index| {
-            projection_incidence
-                .incidence
-                .unknown_names
-                .get(*unknown_index)
-                .and_then(|unknown| projection_y_index(unknown, projection_incidence))
-        })
-        .collect()
-}
-
-fn validate_complete_algebraic_projection_plan(
-    blocks: &[solve::AlgebraicProjectionBlock],
-    rows: &[Vec<solve::LinearOp>],
-    expected_rows: &[usize],
-    projection_incidence: &ProjectionIncidence,
-    context_span: rumoca_core::Span,
-) -> Result<(), LowerError> {
-    let covered_rows = blocks
-        .iter()
-        .flat_map(|block| block.rows.iter().copied())
-        .collect::<BTreeSet<_>>();
-    for row in expected_rows {
-        if covered_rows.contains(row) {
-            continue;
-        }
-        let has_projection_incidence = projection_incidence
-            .incidence
-            .equation_refs
-            .iter()
-            .any(|equation| equation.0 == *row);
-        if has_projection_incidence || statically_nonzero_projection_row(&rows[*row]).is_some() {
-            return Err(lower_contract_violation(
-                format!("algebraic projection plan omits implicit row {row}"),
-                context_span,
-            ));
-        }
-    }
-
-    let covered_y_indices = blocks
-        .iter()
-        .flat_map(|block| block.y_indices.iter().copied())
-        .collect::<BTreeSet<_>>();
-    for y_index in &projection_incidence.unknown_y_indices {
-        if !covered_y_indices.contains(y_index) {
-            return Err(lower_contract_violation(
-                format!("algebraic projection plan omits residual target y[{y_index}]"),
-                context_span,
-            ));
-        }
-    }
-    Ok(())
-}
-
-fn statically_nonzero_projection_row(row: &[solve::LinearOp]) -> Option<f64> {
-    if !row.iter().all(|op| {
-        matches!(
-            op,
-            solve::LinearOp::Const { .. }
-                | solve::LinearOp::Move { .. }
-                | solve::LinearOp::Unary { .. }
-                | solve::LinearOp::Binary { .. }
-                | solve::LinearOp::Compare { .. }
-                | solve::LinearOp::Select { .. }
-                | solve::LinearOp::StoreOutput { .. }
-        )
-    }) {
-        return None;
-    }
-    let value = rumoca_eval_solve::eval_row_with_context(
-        row,
-        &[],
-        &[],
-        0.0,
-        rumoca_eval_solve::RowEvalContext::default(),
-    )
-    .ok()?;
-    (value.is_finite() && value != 0.0).then_some(value)
-}
-
-fn collect_algebraic_y_indices_for_row(
-    row: &[solve::LinearOp],
-    projection_set: &BTreeSet<usize>,
-) -> BTreeSet<usize> {
-    let mut defs = BTreeMap::<solve::Reg, RowDefUse>::new();
-    let mut outputs = Vec::new();
-    for op in row {
-        match row_def_use(op) {
-            RowDefUseOp::Def { dst, def_use } => {
-                defs.insert(dst, def_use);
-            }
-            RowDefUseOp::Store { src } => outputs.push(src),
-        }
-    }
-    let mut y_indices = BTreeSet::new();
-    let mut visited = BTreeSet::new();
-    let mut stack = outputs;
-    while let Some(reg) = stack.pop() {
-        if !visited.insert(reg) {
-            continue;
-        }
-        let Some(def_use) = defs.get(&reg) else {
-            continue;
-        };
-        if let Some(index) = def_use.loaded_y
-            && projection_set.contains(&index)
-        {
-            y_indices.insert(index);
-        }
-        stack.extend(def_use.inputs.iter().copied());
-    }
-    y_indices
-}
-
-#[derive(Debug)]
-struct RowDefUse {
-    loaded_y: Option<usize>,
-    inputs: Vec<solve::Reg>,
-}
-
-enum RowDefUseOp {
-    Def { dst: solve::Reg, def_use: RowDefUse },
-    Store { src: solve::Reg },
-}
-
-fn row_def_use(op: &solve::LinearOp) -> RowDefUseOp {
-    use solve::LinearOp as Op;
-    match *op {
-        Op::Const { dst, .. } | Op::LoadTime { dst } | Op::LoadP { dst, .. } => {
-            def_use(dst, None, Vec::new())
-        }
-        Op::LoadY { dst, index } => def_use(dst, Some(index), Vec::new()),
-        Op::LoadSeed { dst, .. } => def_use(dst, None, Vec::new()),
-        Op::LoadIndexedP { dst, index, .. } | Op::LoadIndexedSeed { dst, index, .. } => {
-            def_use(dst, None, vec![index])
-        }
-        Op::Move { dst, src } | Op::Unary { dst, arg: src, .. } => def_use(dst, None, vec![src]),
-        Op::Binary { dst, lhs, rhs, .. } | Op::Compare { dst, lhs, rhs, .. } => {
-            def_use(dst, None, vec![lhs, rhs])
-        }
-        Op::Select {
-            dst,
-            cond,
-            if_true,
-            if_false,
-        } => def_use(dst, None, vec![cond, if_true, if_false]),
-        Op::LinearSolveComponent {
-            dst,
-            matrix_start,
-            rhs_start,
-            n,
-            ..
-        } => def_use(
-            dst,
-            None,
-            reg_range(matrix_start, n * n)
-                .chain(reg_range(rhs_start, n))
-                .collect(),
-        ),
-        Op::TableBounds { dst, table_id, .. } => def_use(dst, None, vec![table_id]),
-        Op::TableLookup {
-            dst,
-            table_id,
-            column,
-            input,
-        }
-        | Op::TableLookupSlope {
-            dst,
-            table_id,
-            column,
-            input,
-        } => def_use(dst, None, vec![table_id, column, input]),
-        Op::TableNextEvent {
-            dst,
-            table_id,
-            time,
-        } => def_use(dst, None, vec![table_id, time]),
-        Op::RandomInitialState {
-            dst,
-            local_seed,
-            global_seed,
-            ..
-        } => def_use(dst, None, vec![local_seed, global_seed]),
-        Op::RandomResult {
-            dst,
-            state_start,
-            state_len,
-            ..
-        }
-        | Op::RandomState {
-            dst,
-            state_start,
-            state_len,
-            ..
-        } => def_use(dst, None, reg_range(state_start, state_len).collect()),
-        Op::ImpureRandomInit { dst, seed } => def_use(dst, None, vec![seed]),
-        Op::ImpureRandom { dst, id, .. } => def_use(dst, None, vec![id]),
-        Op::ImpureRandomInteger {
-            dst,
-            id,
-            imin,
-            imax,
-            ..
-        } => def_use(dst, None, vec![id, imin, imax]),
-        Op::ExternalCall {
-            dst,
-            args,
-            arg_count,
-            ..
-        } => def_use(dst, None, args.into_iter().take(arg_count).collect()),
-        Op::StoreOutput { src } => RowDefUseOp::Store { src },
-    }
-}
-
-fn def_use(dst: solve::Reg, loaded_y: Option<usize>, inputs: Vec<solve::Reg>) -> RowDefUseOp {
-    RowDefUseOp::Def {
-        dst,
-        def_use: RowDefUse { loaded_y, inputs },
-    }
-}
-
-fn reg_range(start: solve::Reg, len: usize) -> impl Iterator<Item = solve::Reg> {
-    (0..len).filter_map(move |offset| start.checked_add(offset.try_into().ok()?))
-}
-
-struct ProjectionIncidence {
-    incidence: Incidence,
-    unknown_y_indices: Vec<usize>,
-}
-
-fn algebraic_projection_incidence(
-    row_to_vars: &BTreeMap<usize, BTreeSet<usize>>,
-    context_span: rumoca_core::Span,
-) -> Result<ProjectionIncidence, LowerError> {
-    let unknown_y_set = row_to_vars
-        .values()
-        .flat_map(|vars| vars.iter().copied())
-        .collect::<BTreeSet<_>>();
-    let mut unknown_y_indices = lower_vec_with_capacity(
-        unknown_y_set.len(),
-        "projection unknown index count",
-        context_span,
-    )?;
-    unknown_y_indices.extend(unknown_y_set);
-
-    let mut unknown_names = lower_vec_with_capacity(
-        unknown_y_indices.len(),
-        "projection unknown name count",
-        context_span,
-    )?;
-    for y_idx in &unknown_y_indices {
-        unknown_names.push(projection_unknown_id(*y_idx));
-    }
-
-    let unknown_positions = unknown_y_indices
-        .iter()
-        .copied()
-        .enumerate()
-        .map(|(local_idx, y_idx)| (y_idx, local_idx))
-        .collect::<BTreeMap<_, _>>();
-
-    let mut equation_refs = lower_vec_with_capacity(
-        row_to_vars.len(),
-        "projection equation ref count",
-        context_span,
-    )?;
-    let mut eq_unknowns = lower_vec_with_capacity(
-        row_to_vars.len(),
-        "projection equation unknown count",
-        context_span,
-    )?;
-    for (row_idx, vars) in row_to_vars {
-        equation_refs.push(EquationRef(*row_idx));
-        let mut unknowns =
-            lower_hash_set_with_capacity(vars.len(), "projection row unknown count", context_span)?;
-        for y_idx in vars {
-            if let Some(local_idx) = unknown_positions.get(y_idx).copied() {
-                unknowns.insert(local_idx);
-            }
-        }
-        eq_unknowns.push(unknowns);
-    }
-
-    Ok(ProjectionIncidence {
-        incidence: Incidence::new(eq_unknowns, equation_refs, unknown_names),
-        unknown_y_indices,
-    })
-}
-
-fn projection_unknown_id(y_idx: usize) -> UnknownId {
-    UnknownId::SolverY(y_idx)
-}
-
-fn projection_y_index(
-    unknown: &UnknownId,
-    projection_incidence: &ProjectionIncidence,
-) -> Option<usize> {
-    projection_incidence
-        .incidence
-        .unknown_names
-        .iter()
-        .position(|candidate| candidate == unknown)
-        .and_then(|idx| projection_incidence.unknown_y_indices.get(idx).copied())
-}
-
-fn lower_blt_projection_blocks(
-    blocks: &[BltBlock],
-    row_targets: &[Option<solve::ScalarSlot>],
-    projection_incidence: &ProjectionIncidence,
-    context_span: rumoca_core::Span,
-) -> Result<Vec<solve::AlgebraicProjectionBlock>, LowerError> {
-    let mut lowered = lower_vec_with_capacity(
-        blocks.len(),
-        "algebraic projection block count",
-        context_span,
-    )?;
-    for block in blocks {
-        let block = match block {
-            BltBlock::Scalar { equation, unknown } => {
-                projection_y_index(unknown, projection_incidence)
-                    .map(|y_index| {
-                        scalar_projection_block(
-                            equation.0,
-                            y_index,
-                            row_targets,
-                            projection_incidence,
-                            context_span,
-                        )
-                    })
-                    .transpose()?
-            }
-            BltBlock::AlgebraicLoop {
-                equations,
-                unknowns,
-            } => lower_algebraic_loop_projection_block(
-                equations,
-                unknowns,
-                row_targets,
-                projection_incidence,
-                context_span,
-            )?,
-        };
-        if let Some(block) = block {
-            lowered.push(block);
-        }
-    }
-    merge_overlapping_projection_blocks(lowered, context_span)
-}
-
-fn merge_overlapping_projection_blocks(
-    blocks: Vec<solve::AlgebraicProjectionBlock>,
-    context_span: rumoca_core::Span,
-) -> Result<Vec<solve::AlgebraicProjectionBlock>, LowerError> {
-    let mut merged = lower_vec_with_capacity(
-        blocks.len(),
-        "merged algebraic projection block count",
-        context_span,
-    )?;
-    for block in blocks {
-        merge_projection_block(&mut merged, block, context_span)?;
-    }
-    Ok(merged)
-}
-
-fn merge_projection_block(
-    merged: &mut Vec<solve::AlgebraicProjectionBlock>,
-    mut block: solve::AlgebraicProjectionBlock,
-    context_span: rumoca_core::Span,
-) -> Result<(), LowerError> {
-    let mut idx = 0;
-    while idx < merged.len() {
-        if projection_blocks_overlap(&merged[idx], &block) {
-            let previous = merged.remove(idx);
-            block = combine_projection_blocks(previous, block, context_span)?;
-            idx = 0;
-        } else {
-            idx += 1;
-        }
-    }
-    merged.push(block);
-    Ok(())
-}
-
-fn projection_blocks_overlap(
-    lhs: &solve::AlgebraicProjectionBlock,
-    rhs: &solve::AlgebraicProjectionBlock,
-) -> bool {
-    lhs.y_indices
-        .iter()
-        .any(|index| rhs.y_indices.binary_search(index).is_ok())
-}
-
-fn combine_projection_blocks(
-    lhs: solve::AlgebraicProjectionBlock,
-    rhs: solve::AlgebraicProjectionBlock,
-    context_span: rumoca_core::Span,
-) -> Result<solve::AlgebraicProjectionBlock, LowerError> {
-    let causal_step_count = lhs
-        .causal_steps
-        .len()
-        .checked_add(rhs.causal_steps.len())
-        .ok_or_else(|| {
-            lower_contract_violation(
-                "merged algebraic projection causal-step count overflows host index range"
-                    .to_string(),
-                context_span,
-            )
-        })?;
-    let mut causal_steps = lower_vec_with_capacity(
-        causal_step_count,
-        "merged algebraic projection causal-step count",
-        context_span,
-    )?;
-    causal_steps.extend(lhs.causal_steps);
-    causal_steps.extend(rhs.causal_steps);
-    Ok(solve::AlgebraicProjectionBlock {
-        rows: merge_unique(
-            lhs.rows,
-            rhs.rows,
-            "merged algebraic projection row count",
-            context_span,
-        )?,
-        y_indices: merge_unique(
-            lhs.y_indices,
-            rhs.y_indices,
-            "merged algebraic projection target count",
-            context_span,
-        )?,
-        causal_steps,
-    })
-}
-
-fn merge_unique(
-    lhs: Vec<usize>,
-    rhs: Vec<usize>,
-    context: &'static str,
-    context_span: rumoca_core::Span,
-) -> Result<Vec<usize>, LowerError> {
-    let capacity = lhs.len().checked_add(rhs.len()).ok_or_else(|| {
-        lower_contract_violation(
-            format!("{context} overflows host index range"),
-            context_span,
-        )
-    })?;
-    let mut merged = lower_vec_with_capacity(capacity, context, context_span)?;
-    merged.extend(lhs);
-    merged.extend(rhs);
-    merged.sort_unstable();
-    merged.dedup();
-    Ok(merged)
-}
-
-fn scalar_projection_block(
-    row: usize,
-    y_index: usize,
-    row_targets: &[Option<solve::ScalarSlot>],
-    projection_incidence: &ProjectionIncidence,
-    context_span: rumoca_core::Span,
-) -> Result<solve::AlgebraicProjectionBlock, LowerError> {
-    let mut rows = lower_vec_with_capacity(
-        1,
-        "scalar algebraic projection block row count",
-        context_span,
-    )?;
-    rows.push(row);
-    let mut target_set = BTreeSet::from([y_index]);
-    if let Some(solve::ScalarSlot::Y { index, .. }) = row_targets.get(row).copied().flatten()
-        && projection_incidence.unknown_y_indices.contains(&index)
-    {
-        target_set.insert(index);
-    }
-    let mut y_indices = lower_vec_with_capacity(
-        target_set.len(),
-        "scalar algebraic projection block target count",
-        context_span,
-    )?;
-    y_indices.extend(target_set);
-    let causal_target = row_targets
-        .get(row)
-        .copied()
-        .flatten()
-        .and_then(|target| match target {
-            solve::ScalarSlot::Y { index, .. } => Some(index),
-            _ => None,
-        })
-        .filter(|target| y_indices.contains(target));
-    let causal_steps = if let Some(target) = causal_target {
-        vec![solve::AlgebraicProjectionStep {
-            row,
-            y_index: target,
-        }]
-    } else {
-        Vec::new()
-    };
-    Ok(solve::AlgebraicProjectionBlock {
-        rows,
-        y_indices,
-        causal_steps,
-    })
-}
-
-fn sorted_set_values(
-    values: BTreeSet<usize>,
-    context: &'static str,
-    context_span: rumoca_core::Span,
-) -> Result<Vec<usize>, LowerError> {
-    let mut out = lower_vec_with_capacity(values.len(), context, context_span)?;
-    out.extend(values);
-    Ok(out)
-}
-
-fn collect_equation_rows(
-    equations: &[EquationRef],
-    context_span: rumoca_core::Span,
-) -> Result<Vec<usize>, LowerError> {
-    let mut rows = lower_vec_with_capacity(
-        equations.len(),
-        "algebraic loop projection row count",
-        context_span,
-    )?;
-    for equation in equations {
-        rows.push(equation.0);
-    }
-    Ok(rows)
-}
-
-fn lower_algebraic_loop_projection_block(
-    equations: &[EquationRef],
-    unknowns: &[UnknownId],
-    row_targets: &[Option<solve::ScalarSlot>],
-    projection_incidence: &ProjectionIncidence,
-    context_span: rumoca_core::Span,
-) -> Result<Option<solve::AlgebraicProjectionBlock>, LowerError> {
-    let rows = collect_equation_rows(equations, context_span)?;
-    let y_indices = sorted_set_values(
-        loop_projection_target_set(unknowns, row_targets, &rows, projection_incidence),
-        "algebraic loop projection target count",
-        context_span,
-    )?;
-    if rows.is_empty() || y_indices.is_empty() {
-        return Ok(None);
-    }
-    Ok(Some(solve::AlgebraicProjectionBlock {
-        rows,
-        y_indices,
-        causal_steps: Vec::new(),
-    }))
-}
-
-fn loop_projection_target_set(
-    unknowns: &[UnknownId],
-    row_targets: &[Option<solve::ScalarSlot>],
-    rows: &[usize],
-    projection_incidence: &ProjectionIncidence,
-) -> BTreeSet<usize> {
-    let mut y_indices = BTreeSet::new();
-    for unknown in unknowns {
-        if let Some(index) = projection_y_index(unknown, projection_incidence) {
-            y_indices.insert(index);
-        }
-    }
-    for row in rows {
-        if let Some(solve::ScalarSlot::Y { index, .. }) = row_targets.get(*row).copied().flatten()
-            && projection_incidence.unknown_y_indices.contains(&index)
-        {
-            y_indices.insert(index);
-        }
-    }
-    y_indices
 }
 
 pub fn solver_vector_names(
